add x-goog-user-project header and custom user-agent
x-goog-user-project allows to set the billing/quota project governing
the API calls, and will be by default set to the project that is being
inspected (instead of the gcp-doctor project). b/193027624, b/197926813

This required some refactoring in the get_api stubs to make the tests pass
(and remove obsolete code).

Also, set a user-agent string to making it clear who is doing these
requests.

Change-Id: I595d3ce7f22423ae480c9b2fb2d6ae75609f3189
GitOrigin-RevId: e459cd2ff00f94fd97c94cf0c39956caed9a3a3c