Rename gcp-doctor to gcpdiag (b/200636953)
This commit only renames directories/files and any required code changes
to make the code work after these renames. Further code changes will be
done in other commits.

Note also that we need to move the main script to bin, because we can't
have both the module directory 'gcpdiag' and the binary 'gcpdiag' in the
root directory. This was previously possible because of gcp-doctor vs
gcp_doctor.

Change-Id: I154fedf5b7f64b1056e6564713685ae68e16ed84
GitOrigin-RevId: d44171b112cbe7d5ff2a5cea247fa873354a3dfd