Fix checking Cloud Build service account role rule.
Ref: https://cloud.google.com/build/docs/cloud-build-service-account

Cloud Build service account is only used to create builds by triggers.
If user is creating builds manually it does not need to have any role.
It is also used as default SA used during a build, but this usage does
not require it to have any role.

In addition it is not editor but "Cloud Build Service Account role that
is granted to this SA by default.

Change-Id: I4214d6c0148829b2eacf52fc588ffbffb4c3e877
GitOrigin-RevId: 521ef902d1910606e1d8a24dd45caca06ffce01c