Publish website changes immediately, not only on release.
We want changes to propagate as quickly as possible to the website,
so that for example rule description changes are immediately appearing
and don't need to wait for a release.

Use staging.gcpdiag.dev for presubmit. This is not ideal because
any other presubmit job will overwrite the changes, but maybe later
it can be replaced with a better model (like for example by using
URLs like staging.gcpdiag.dev/N/..., where N is a rotating number).

Also fix the job to use the correct service account (the default
one from the kokoro GCE instance).

Change-Id: Ieb91f7403c397b7ce98ae9f32c01bca1cc6ebecc