Merge "Improve org policy query performance bug: 433558560" -- Branch commit log -- commit ac5cd4ea807b4320df9a8107af671ba4ec5a5278 Author:  Joel Jacob <joeljacob@google.com> Date:    2025-07-22T17:40:42-04:00
    Improve org policy query performance
bug: 433558560

Here's the breakdown of the current strategy:
Fast, Cached Pre-fetch (The "Shortcut"):
The function _get_effective_org_policy_all_constraints is designed to
be fast. It runs get_all_project_org_policies which only
fetches policies directly set on the project, ignoring inheritance.
This is quick and the results are cached.

Slower, Precise Fallback: The main function, get_effective_org_policy,
first checks that fast cache. If the specific constraint it's
looking for isn't in the cache (meaning it's not set directly
on the project), it then makes a different, more precise API call:
crm_api.projects().getEffectiveOrgPolicy. This call specifically
requests the effective policy for a single constraint, which
is slower but accurate.

Change-Id: I5ea00ff568cf7dea4dde560e1a93c34ad0c61e79
GitOrigin-RevId: 96672c3421f7d60f3680a96f82f21fa54f068641